+++ opengrok-indexer/src/main/java/org/opengrok/indexer/analysis/LinkageMatchedEvent.java
@@ -33,8 +33,8 @@ public class LinkageMatchedEvent {
     private final Object source;
     private final String str;
     private final LinkageType linkageType;
-    private final int start;
-    private final int end;
+    private final long start;
+    private final long end;
     private final String lstr;
 
     /**
@@ -47,7 +47,7 @@ public class LinkageMatchedEvent {
      * @param end the text end position
      */
     public LinkageMatchedEvent(Object source, String str,
-        LinkageType linkageType, int start, int end) {
+        LinkageType linkageType, long start, long end) {
         this.source = source;
         this.str = str;
         this.linkageType = linkageType;
@@ -66,7 +66,7 @@ public class LinkageMatchedEvent {
      * @param lstr the text link string
      */
     public LinkageMatchedEvent(Object source, String str,
-        LinkageType linkageType, int start, int end, String lstr) {
+        LinkageType linkageType, long start, long end, String lstr) {
         this.source = source;
         this.str = str;
         this.linkageType = linkageType;
@@ -112,7 +112,7 @@ public class LinkageMatchedEvent {
      * Gets the text start position.
      * @return the initial value
      */
-    public int getStart() {
+    public long getStart() {
         return start;
     }
 
@@ -120,7 +120,7 @@ public class LinkageMatchedEvent {
      * Gets the text end position.
      * @return the initial value
      */
-    public int getEnd() {
+    public long getEnd() {
         return end;
     }
 }
